\name{drawTrackOverlay-methods}
\docType{methods}
\alias{drawTrackOverlay-methods}
\alias{drawTrackOverlay,Segmentation-method}
\alias{drawTrackOverlay,Smoothing-method}
\title{This method does the drawing of a track overlay. One should
  implement this method to gain track plotting functionality.}
\description{
  Methods for function \code{drawTrackOverlay} in Package `GenomeGraphs'
}
\section{Methods}{
  \describe{
    \item{Segmentation}{}
    \item{Smoothing}{}
}}
\keyword{methods}
\keyword{classes}
