\name{getGenomicRange}
\docType{methods}
\alias{getGenomicRange}
\alias{getGenomicRange,BaseTrack-method}
\alias{getGenomicRange,ExonArray-method}
\alias{getGenomicRange,gdObject-method}
\alias{getGenomicRange,Gene-method}
\alias{getGenomicRange,GeneRegion-method}
\alias{getGenomicRange,GenericArray-method}
\alias{getGenomicRange,Transcript-method}
\alias{getGenomicRange,TranscriptRegion-method}
\title{Retrieves the genomic range of an object}
\description{
 \code{getGenomicRange} returns the genomic range of an object
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "BaseTrack")}}{
#to be added
}

\item{\code{signature(obj = "ExonArray")}}{
#to be added
}

\item{\code{signature(obj = "gdObject")}}{
#to be added
}

\item{\code{signature(obj = "Gene")}}{
#to be added
}

\item{\code{signature(obj = "GeneRegion")}}{
#to be added
}

\item{\code{signature(obj = "GenericArray")}}{
#to be added
}

\item{\code{signature(obj = "Transcript")}}{
#to be added
}

\item{\code{signature(obj = "TranscriptRegion")}}{
#to be added
}
}}
\keyword{methods}
