\name{getSize}
\alias{getSize}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{gets the size}
\description{
  Gets the size display parameter
}
\usage{
getSize(obj, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{An object, usually a gdObject.}
  \item{...}{Ignored}
}
\examples{
#to be added
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{hplot}

