\name{ideogramTab}
\alias{ideogramTab}
\docType{data}
\title{Contains info to plot ideograms}
\description{
 Contains info to plot ideograms
}
\format{
  The format is:
 chr "ideogramTab"
}
\source{
NCBI
}
\examples{
data(ideogramTab)
## maybe str(ideogramTab) ; plot(ideogramTab) ...
}
\keyword{datasets}
