\name{ImplementsTrackOverlay-class}
\Rdversion{1.1}
\docType{class}
\alias{ImplementsTrackOverlay-class}

\title{Class "ImplementsTrackOverlay"}
\description{The interface to be implemented to overlay tracks.}
\section{Objects from the Class}{
  This object should not be instantiated, but rather this class should
  be extended to implement a particular interface.
}
\section{Slots}{
  \describe{
    \item{\code{trackOverlay}:}{Object of class \code{"TrackOverlayOrNull"}}
  }
}
\section{Methods}{
  No methods defined with class "ImplementsTrackOverlay" in the signature.
}
\examples{
showClass("ImplementsTrackOverlay")
}
\keyword{classes}
