\name{Overlay-class}
\docType{class}
\alias{Overlay-class}

\title{Class "Overlay"}
\description{Superclass of overlay objects.}
\section{Objects from the Class}{
  Objects from this class are generally not created.
}
\section{Slots}{
  \describe{
    \item{\code{dp}:}{Object of class \code{"DisplayPars"}}
  }
}
\section{Extends}{
Class \code{"\linkS4class{gdObject}"}, directly.
}
\section{Methods}{
  No methods defined with class "Overlay" in the signature.
}
\examples{
showClass("Overlay")
}
\keyword{classes}
