\name{TrackOverlay-class}
\Rdversion{1.1}
\docType{class}
\alias{TrackOverlay-class}

\title{Class "TrackOverlay"}
\description{Parent class for track overlay objects, such as
  \code{Smoothing} and \code{Segmentation}}
\section{Objects from the Class}{
  Objects of this class are not instantiated, but rather this class
  should be extended. 
}
\section{Slots}{
  \describe{
    \item{\code{dp}:}{Object of class \code{"DisplayPars"} ~~ }
  }
}
\section{Extends}{
  Class \code{"\linkS4class{gdObject}"}, directly.
  Class \code{"TrackOverlayOrNull"}, directly.
}
\section{Methods}{
  No methods defined with class "TrackOverlay" in the signature.
}
\examples{
showClass("TrackOverlay")
}
\keyword{classes}
