\name{unrData}
\alias{unrData}
\docType{data}
\title{Contains exon array data}
\description{
 Contains exon array data from the publically available dataset on human tissue panels, given by Affymetrix.  The data was contributed to the package by Elizabeth Purdom.

}

\examples{
#
}

\keyword{datasets}
