\name{unrNProbes}
\alias{unrNProbes}
\docType{data}
\title{Contains exon array data}
\description{
 Contains the number of probes per exon array probeset id from the publically available dataset on human tissue panels, given by Affymetrix.  The data was contributed to the package by Elizabeth Purdom.

}

\examples{
#
}

\keyword{datasets}
