\name{unrPositions}
\alias{unrPositions}
\docType{data}
\title{Contains probe start and end positions of exon array probes}
\description{
 Contains probe start and end positions from the publically available dataset on human tissue panels, given by Affymetrix.  The data was contributed to the package by Elizabeth Purdom.

}

\examples{
#
}

\keyword{datasets}
