\name{seqDataEx}
\alias{seqDataEx}
\docType{data}
\title{This is an example data set from chromosome 4 of yeast from
  various publicly available datasets.}
\description{
  This was a small dataset constructed from publicly available
  datasets. Please see references for details.
}
\usage{data(seqDataEx)}
\format{
  data("seqDataEx", package = "GenomeGraphs")
  names(seqDataEx)
}
\source{
}
\references{
  Ugrappa Nagalakshmi et. al. The transcriptional landscape of the yeast genome defined by RNA sequencing.
  Science, 2008

  Lior David et. al. A high-resolution map of transcription in the yeast genome.
  Proc Natl Acad Sci U S A, (2006)
  
  William Lee A high-resolution atlas of nucleosome occupancy in yeast.
  Nat Genet, 2007
  
  Adam Siepel, et. al. Evolutionarily conserved elements in vertebrate, insect, worm, and yeast genomes.
  Genome Res, 2005
}
\examples{
data(seqDataEx)
}
\keyword{datasets}
